# api/api.yaml
openapi: "3.0.0"
info:
  title: "Nuts Demo EHR API"
  version: "1.0.0"
paths:
  /auth:
    post:
      operationId: setCustomer
      description: sets the current customer ID in the JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Customer"
      responses:
        "200":
          description: A session was succesfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionToken"
        "403":
          description: Invalid credentials

  # Password authentication
  /auth/passwd:
    post:
      description: Authenticate using password credential. If successful an authenticated session is created and returned.
      operationId: authenticateWithPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordAuthenticateRequest"
      responses:
        "204":
          description: Authentication successful.
        "403":
          description: Invalid credentials
  # IRMA authentication
...
  /auth/irma/session/{sessionToken}/result:
    parameters:
      - name: sessionToken
        in: path
        description: IRMA session ID
        required: true
        schema:
          type: string
    get:
      description: |
        After a successful IRMA signing session the resulting signature can be fetched using his endpoint.
      operationId: getIRMAAuthenticationResult
      responses:
        "200":
          description: Session result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionToken"
...
  /auth/dummy/session/{sessionToken}/result:
    parameters:
      - name: sessionToken
        in: path
        description: Dummy session ID
        required: true
        schema:
          type: string
    get:
      description: |
        After a successful Dummy signing session the resulting token can be fetched using his endpoint.
      operationId: getDummyAuthenticationResult
      responses:
        "200":
          description: Session result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionToken"
  /private:
    get:
      description: Checks whether the current session is valid. If not, the client should authenticate before calling other API operations.
      operationId: checkSession
      responses:
        "204":
          description: The session is valid.
        "400":
          description: The session is invalid.
...
  /private/reports/{patientID}:
    parameters:
      - name: patientID
        in: path
        description: The patient ID
        required: true
        schema:
          type: string
    get:
      parameters:
        - name: episodeID
          in: query
          description: The identifier of episode the report must be part of.
          required: false
          schema:
            type: string
      description: Get list of reports for a patient
      operationId: getReports
      responses:
        200:
          description: The list of reports for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      description: Create a new record for a patient
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Report"
      responses:
        200:
          description: The creation of the new report is completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /private/prescriptions/{patientID}:
    parameters:
      - name: patientID
        in: path
        description: The patient ID
        required: true
        schema:
          type: string
    get:
      parameters:
        - name: episodeID
          in: query
          description: The identifier of episode the prescription must be part of.
          required: false
          schema:
            type: string
      description: Get list of prescriptions for a patient
      operationId: getPrescriptions
      responses:
        200:
          description: The list of prescriptions for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"

    post:
      description: Create a new prescription for a patient
      operationId: createPrescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Prescription"
      responses:
        200:
          description: The creation of the new prescription is completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /private/medications:
    post:
      description: Create a medication
      operationId: createMedication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Medication"
      responses:
        200:
          description: The created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        404:
          description: The patient or the doctor does not exist
  /private/medication/{medicationID}:
    get:
      description: Get the medication by the medicationID
      operationId: getMedication
      parameters:
        - name: medicationID
          in: path
          description: The medication ID
          required: true
          schema:
            type: string
      responses:
        200:
          description: The requested medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        404:
          description: The medication does not exist
...
components:
  schemas:
    Customer:
      type: object
      description: A customer object.
      required:
        - id
        - name
        - active
      properties:
        id:
          type: integer
          description: The internal customer ID.
          example: 1
        did:
          type: string
          description: The customer DID.
          example: "did:nuts:123"
        name:
          type: string
          description: Internal name for this customer.
        city:
          type: string
          description: Locality for this customer.
        domain:
          type: string
          description: The email domain of the care providers employees, required for logging in.
        active:
          type: boolean
          description: If a VC has been issued for this customer.
...
    ObjectID:
      description: An internal object UUID which can be used as unique identifier for entities.
      type: string
      example:
        - 665d6390-4823-4829-a3bf-4257e0b6799
      pattern: "/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i"
...
    SessionToken:
      type: object
      description: Result of a signing session.
      required:
        - token
      properties:
        token:
          type: string
          description: the result from a signing session. It's an updated JWT.
    PasswordAuthenticateRequest:
      required:
        - customerID
        - password
      properties:
        customerID:
          description: Internal ID of the customer for which is being logged in
          type: integer
          example: 1
        password:
          type: string
...
    Report:
      required:
        - id
        - patientID
        - type
        - value
        - source
      properties:
        id:
          $ref: "#/components/schemas/ObjectID"
        patientID:
          $ref: "#/components/schemas/ObjectID"
        episodeID:
          $ref: "#/components/schemas/ObjectID"
        type:
          type: string
        value:
          type: string
        source:
          type: string
        episodeName:
          type: string
    Prescription:
      required:
        - id
        - patientID
      properties:
        id:
          $ref: "#/components/schemas/ObjectID"
        patientID:
          $ref: "#/components/schemas/ObjectID"
        episodeID:
          $ref: "#/components/schemas/ObjectID"
        medicationID:
          $ref: "#/components/schemas/ObjectID"
        medicationName:
          type: string
        source:
          type: string
        episodeName:
          type: string
    Medication:
      required:
        - id
        - name
      properties:
        id:
          $ref: "#/components/schemas/ObjectID"
        name:
          type: string
        source:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] # Apply Bearer Auth to all endpoints
servers:
  - url: "/web"